Recompile tokens in the C++ side before returning it to R
