+++ R/tokens_ngrams.R
@@ -134,7 +134,8 @@ tokens_ngrams.tokens <- function(x, n = 2L, skip = 0L, concatenator = "_") {
     attr(x, "ngrams") <- as.integer(n)
     attr(x, "skip") <- as.integer(skip)
     attr(x, "concatenator") <- concatenator
-    tokens_hashed_recompile(x)
+    Encoding(types(x)) <- "UTF-8"
+    return(x)
 }
