+++ src/tokens_ngrams_mt.cpp
@@ -1,4 +1,5 @@
 //#include "dev.h"
+#include "recompile.h"
 #include "quanteda.h"
 using namespace quanteda;
 
@@ -218,12 +219,7 @@ List qatd_cpp_tokens_ngrams(const List texts_,
     }
 #endif
     //dev::stop_timer("Token generation", timer);
-    
-    // Return IDs as attribute
-    ListOf<IntegerVector> output_ = Rcpp::wrap(output);
-    output_.attr("types") = types_ngram;
-    return output_;
-    
+    return recompile(output, types_ngram);
 }
